Former-commit-id: 601449e26c24b9edd3902806f7ae7fbca89b2350

+++ build/docker-compose-down.sh
@@ -19,6 +19,7 @@ if [ -n "${DAPP}" ]; then
 
 fi
 
+echo "=========== # down docker-compose ============="
 echo "=========== # env setting ============="
 echo "DAPP=$DAPP"
 echo "COMPOSE_FILE=$COMPOSE_FILE"
@@ -29,12 +30,16 @@ echo "COMPOSE_PROJECT_NAME=$COMPOSE_PROJECT_NAME"
 function down() {
     echo "=========== # docker-compose ps ============="
     docker-compose ps
-
-    # remove exsit container
-    echo "=========== # docker-compose down ============="
-    docker-compose down --rmi local
-
-    docker-compose ps
+    # shellchk not recommend the first way
+    # remains=( $(docker-compose ps -q | awk '{print $1}') )
+    mapfile -t remains < <(docker-compose ps -q | awk '{print $1}')
+    num=${#remains[@]}
+    echo "container num=$num"
+    if [ "$num" -gt 0 ]; then
+        # remove exsit container
+        echo "=========== # docker-compose down ============="
+        docker-compose down --rmi local
+    fi
 
 }
